Refactoring, removal of button subclasses.
